requirements.txt
# #aws-cdk.aws-s3==1.100.0
# #aws-cdk.aws-s3-deployment==1.100.0
# aws-cdk.aws-dynamodb==1.100.0
# aws-cdk.aws-lambda==1.100.0
# aws-cdk.aws-apigateway==1.100.0
# boto3
# requests
# httpie

aws-cdk-lib~=2.0
constructs>=10.0.0,<11.0.0
boto3
requests